separation of stats utils into different files. Comments
